refactor(pagination): move logic into custom hook
